# Speed up processing_time_case function