--- a/docs/pkgdown.js
+++ b/docs/pkgdown.js
@@ -30,10 +30,10 @@
   /* Clipboard --------------------------*/
 
   function changeTooltipMessage(element, msg) {
-    var tooltipOriginalTitle=element.getAttribute('data-original-title');
-    element.setAttribute('data-original-title', msg);
+    var tooltipOriginalTitle=element.getAttribute('data-bs-original-title');
+    element.setAttribute('data-bs-original-title', msg);
     $(element).tooltip('show');
-    element.setAttribute('data-original-title', tooltipOriginalTitle);
+    element.setAttribute('data-bs-original-title', tooltipOriginalTitle);
   }
 
   if(ClipboardJS.isSupported()) {
@@ -60,7 +60,7 @@
         e.clearSelection();
       });
 
-      clipboard.on('error', function() {
+      clipboard.on('error', function(e) {
         changeTooltipMessage(e.trigger,'Press Ctrl+C or Command+C to copy');
       });
 
@@ -153,4 +153,10 @@ async function searchFuse(query, callback) {
   });
 })(window.jQuery || window.$)
 
-
+document.addEventListener('keydown', function(event) {
+  // Check if the pressed key is '/'
+  if (event.key === '/') {
+    event.preventDefault();  // Prevent any default action associated with the '/' key
+    document.getElementById('search-input').focus();  // Set focus to the search input
+  }
+});