--- a/man/edeaR.Rd
+++ b/man/edeaR.Rd
@@ -2,8 +2,8 @@
 % Please edit documentation in R/edeaR.R
 \docType{package}
 \name{edeaR}
-\alias{edeaR}
 \alias{edeaR-package}
+\alias{edeaR}
 \title{edeaR - Exploratory and Descriptive Event-based data Analysis in R}
 \description{
 This package provides several useful techniques for Exploratory and Descriptive analysis of event based data in R, developed by the Business Informatics Research Group of Hasselt University.