--- a/man/filter_trim_lifecycle.Rd
+++ b/man/filter_trim_lifecycle.Rd
@@ -79,30 +79,30 @@ Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence T
 \code{\link[bupaR]{lifecycle_id}}
 
 Other filters: 
+\code{\link{filter_activity}()},
 \code{\link{filter_activity_frequency}()},
 \code{\link{filter_activity_instance}()},
 \code{\link{filter_activity_presence}()},
-\code{\link{filter_activity}()},
-\code{\link{filter_case_condition}()},
 \code{\link{filter_case}()},
-\code{\link{filter_endpoints_condition}()},
+\code{\link{filter_case_condition}()},
 \code{\link{filter_endpoints}()},
+\code{\link{filter_endpoints_condition}()},
 \code{\link{filter_flow_time}()},
 \code{\link{filter_idle_time}()},
 \code{\link{filter_infrequent_flows}()},
-\code{\link{filter_lifecycle_presence}()},
 \code{\link{filter_lifecycle}()},
+\code{\link{filter_lifecycle_presence}()},
+\code{\link{filter_precedence}()},
 \code{\link{filter_precedence_condition}()},
 \code{\link{filter_precedence_resource}()},
-\code{\link{filter_precedence}()},
 \code{\link{filter_processing_time}()},
-\code{\link{filter_resource_frequency}()},
 \code{\link{filter_resource}()},
+\code{\link{filter_resource_frequency}()},
 \code{\link{filter_throughput_time}()},
 \code{\link{filter_time_period}()},
+\code{\link{filter_trace}()},
 \code{\link{filter_trace_frequency}()},
 \code{\link{filter_trace_length}()},
-\code{\link{filter_trace}()},
 \code{\link{filter_trim}()}
 }
 \concept{filters}