… NAMESPACE
… NEWS.md
… README.md
… docs/404.html
… docs/LICENSE-text.html
… docs/articles/filters.html
… docs/articles/index.html
… docs/articles/metrics.html
… docs/articles/metrics_files/figure-html/unnamed-chunk-13-1.png
… docs/articles/metrics_files/figure-html/unnamed-chunk-3-1.png
… docs/articles/metrics_files/figure-html/unnamed-chunk-7-1.png
… docs/articles/metrics_files/figure-html/unnamed-chunk-9-1.png
… docs/articles/queue.html
… docs/articles/queue_files/figure-html/unnamed-chunk-10-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-11-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-12-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-4-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-5-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-6-1.png
… docs/articles/queue_files/figure-html/unnamed-chunk-8-1.png
… docs/articles/work_schedules.html
… docs/authors.html
… docs/index.html
… docs/pkgdown.yml
… docs/reference/activity_frequency.html
… docs/reference/activity_presence.html
… docs/reference/add_fixed_holiday.html
… docs/reference/add_floating_holiday.html
… docs/reference/add_holiday_periods.html
… docs/reference/augment.html