I rewrote parts of the `processing_time_case` function using data.table to speed it up.

The new version takes ~90% less time on my custom dataset (reduced median time of 159 seconds to 12 seconds, 5 evals).
The custom dataset contains almost 200,000 rows. 

On the default datasets (only `processing_time_case`, run on a different device, 50 evals). 
* hospital takes ~47% less time (reduced from ~287ms to ~151ms)
* patients takes ~90% less time
* sepsis takes ~90% less time
* hospital_billing takes ~97% less time
* traffic_fines takes ~98% less time (reduced from ~1600ms to 29ms)

The output is not exactly the same, but I believe that is due to different rounding. I have compared the output from the current and the new version on my custom dataset. When some of the processing_times differ, it is a difference of one (1) in the last/14th decimal. 
Example (underscore added for emphasis):  
- current: 8.8744444444444_**4** days  
- new:       8.8744444444444_**5** days

I think this is acceptable, these are differences on the scale of a [nanosecond](https://www.wolframalpha.com/input?i=1e-14*1+day) , whilst the difference in runtime is significant  for our dataset and usecase.